inserted error checking in accounts_model.php and connections_modle.php.  added error -4 which means account doesnt exist

+++ system/application/models/connections_model.php
@@ -14,47 +14,95 @@ class Connections_model extends Model {
 		$this->load->database();
 	}
 
-	//$inputs is of the form(account_id)
-	//return array of all patients a doctor has
+
+	/**
+	 *  Lists all of the patients a particular doctor has
+	 * 
+	 * @param $inputs
+	 *   Is of the form: array(account_id)
+	 * @return
+	 *   -1 in case of error in a query
+	 * 	  Array with all the patients a doctor has
+	 * 	  NULL if none
+	 * */
 	function list_my_patients($inputs) {
 		$sql = "SELECT *
 			FROM p_d_connection D, patient_account P
 			WHERE (D.patient_id =  P.account_id) AND D.hcp_id = ?";
 		$query = $this->db->query($sql, $inputs);
-		return $query->result_array();
 		
+		if ($this->db->trans_status() === FALSE)
+			return -1;
+			
+		if ($query->num_rows() > 0)
+			return $query->result_array();
+		
+		return NULL;
 	}
 
-	//$inputs is of the form(account_id)
-	//return array of all doctor friends a doctor has
+
+	/**
+	 *  Lists all of the doctor friends a doctor has
+	 * 
+	 * @param $inputs
+	 *   Is of the form: array(account_id)
+	 * @return
+	 *   -1 in case of error in a query
+	 * 	  Array with all the doctor friends
+	 * 	  NULL if none
+	 * */
 	function list_my_colleagues($account_id) {
 		$sql = "SELECT *
 			FROM d_d_connection D, hcp_account H
 			WHERE (D.requester_id = ? AND D.accepter_id = H.account_id)
 			OR    (D.accepter_id = ? AND D.requester_id = H.account_id)";
 		$query = $this->db->query($sql, array($account_id, $account_id));
-		return $query->result_array();
-	
+		
+		if ($this->db->trans_status() === FALSE)
+			return -1;
+			
+		if ($query->num_rows() > 0)	
+			return $query->result_array();
+			
+		return NULL;
 	}	
 
-	//$inputs is of the form(account_id)
-	//returns array of all doctors a pattient has
+
+	/**
+	 *  Lists all of the doctors a patient has
+	 * 
+	 * @param $inputs
+	 *   Is of the form: array(account_id)
+	 * @return
+	 *   -1 in case of error in a query
+	 * 	  Array of all the doctor friends
+	 *    NULL if none
+	 * */
 	function list_my_doctors($inputs) {
 		$sql = "SELECT *
 			FROM p_d_connection D, hcp_account H
 			WHERE (D.patient_id =  ?) AND H.account_id = D.hcp_id";
 		$query = $this->db->query($sql, $inputs);
-		return $query->result_array();
+		
+		if ($this->db->trans_status() === FALSE)
+			return -1;
+			
+		if ($query->num_rows() > 0)	
+			return $query->result_array();
+		
+		return NULL;
 	}
 	
 	
 	/**
-	 * List all pending requests to doctors from a patient
+	 * Lists all pending requests to doctors from a patient
 	 * 
 	 * @param $inputs
 	 *   Is of the form: array(account_id)
-	 * @return
-	 *   Array with all pending requests, NULL otherwise
+	 * @return	
+	 *   -1 in case of error in a query
+	 *    Array with all pending requests
+	 * 	  NULL if none
 	 * */
 	 function pending_todoctors_frompatient($inputs)
 	 {
@@ -62,9 +110,13 @@ class Connections_model extends Model {
 	 		FROM p_d_connection P, hcp_account H
 			WHERE P.patient_id = ? AND AND P.accepted = FALSE AND P.hcp_id = H.account_id";
  		$query = $this->db->query($sql, $inputs);
- 		$result = $query->result_array();
-		if (count($result) >= 1)
-			return $result;
+ 		
+ 		if ($this->db->trans_status() === FALSE)
+			return -1;
+		
+		if ($query->num_rows() > 0)
+			return $query->result_array();
+		
 		return NULL;
 
 	 }
@@ -76,7 +128,9 @@ class Connections_model extends Model {
 	 * @param $inputs
 	 *   Is of the form: array(account_id)
 	 * @return
-	 *   Array with all pending requests, NULL otherwise
+	 *   -1 in case of error in a query
+	 *    Array with all pending requests
+	 * 	  NULL if none
 	 * */
 	 function pending_todoctor_frompatients($inputs)
 	 {
@@ -85,11 +139,16 @@ class Connections_model extends Model {
 			WHERE P.hcp_id = ? AND AND P.accepted = FALSE AND A.account_id = P.patient_id";
  		$query = $this->db->query($sql, $inputs);
  		$result = $query->result_array();
-		if (count($result) >= 1)
-			return $result;
+	
+		if ($this->db->trans_status() === FALSE)
+			return -1;
+		
+		if ($query->num_rows() > 0)
+			return $query->result_array();
+		
 		return NULL;
-
 	 }
+	 
 	 
 	 /**
 	 * List all pending requests to a doctor from doctors
@@ -97,7 +156,9 @@ class Connections_model extends Model {
 	 * @param $inputs
 	 *   Is of the form: array(account_id)
 	 * @return
-	 *   Array with all pending requests, NULL otherwise
+	 * 	 -1 in case of error in a query
+	 *    Array with all pending requests OR 
+	 * 	  NULL if none
 	 * */
 	 function pending_todoctor_fromdoctors($inputs)
 	 {
@@ -106,11 +167,16 @@ class Connections_model extends Model {
 			WHERE D.accepter_id = ? AND AND D.accepted = FALSE AND A.account_id = D.requester_id";
  		$query = $this->db->query($sql, $inputs);
  		$result = $query->result_array();
-		if (count($result) >= 1)
-			return $result;
+		
+		if ($this->db->trans_status() === FALSE)
+			return -1;
+		
+		if ($query->num_rows() > 0)
+			return $query->result_array();
+		
 		return NULL;
-
 	 }
+	 
 	 
 	 /**
 	 * List all pending requests to doctors from a doctor
@@ -118,7 +184,9 @@ class Connections_model extends Model {
 	 * @param $inputs
 	 *   Is of the form: array(account_id)
 	 * @return
-	 *   Array with all pending requests, NULL otherwise
+	 * 	 -1 in case of error in a query
+	 *    Array with all pending requests
+	 * 	  NULL if none	 
 	 * */
 	 function pending_todoctors_fromdoctor($inputs)
 	 {
@@ -126,18 +194,26 @@ class Connections_model extends Model {
 	 		FROM d_d_connection D, hcp_account A
 			WHERE D.requester_id = ? AND AND D.accepted = FALSE AND A.account_id = D.accepter_id";
  		$query = $this->db->query($sql, $inputs);
- 		$result = $query->result_array();
-		if (count($result) >= 1)
-			return $result;
+	
+		if ($this->db->trans_status() === FALSE)
+			return -1;
+		
+		if ($query->num_rows() > 0)
+			return $query->result_array();
+		
 		return NULL;
-
 	 }
 	
 
 	/**
 	 * Tells whether an account is in connection with another account
+	 * 
+	 * @params $inputs
+	 * 	 Is of the form: array(account_id)
 	 * @return
-	 *   TRUE if $a_id is in connection with $b_id, FALSE otherwise
+	 * 	 -1 in case of error in a query
+	 *    TRUE if $a_id is in connection with $b_id
+	 *    FALSE otherwise
 	 * */
 	function is_connected_with($a_id, $b_id) {
 		$sql = "(SELECT *
@@ -155,7 +231,14 @@ class Connections_model extends Model {
 											  $a_id, $b_id,
 											  $a_id, $b_id,
 											  $a_id, $b_id,));
-		return ($query->num_rows() > 0);
+											  
+		if ($this->db->trans_status() === FALSE)
+			return -1;
+			
+		if ($query->num_rows() > 0)
+			return ($query->num_rows() > 0);
+		
+		return FALSE;
 	}
 	
 	
@@ -168,7 +251,7 @@ class Connections_model extends Model {
 	 * @return
 	 *   -1 in case of error in a query
 	 *   -3 if the connection already exists
-	 *   0  if everything goes fine
+	 *    0 if everything goes fine
 	 * 
 	 * @test We still need to test the auto-acceptance if both doctors
 	 * ask for the same connection.
@@ -229,17 +312,16 @@ class Connections_model extends Model {
 	 *
 	 * @return
 	 *   -1 in case of error in a query
-	 *   -2 if the connection does not exist
 	 *   -3 if the connection was already accepted
-	 *   0  if everything goes fine
+	 *    0 if everything goes fine
 	 * 
 	 * */
 	function add_patient_doctor($inputs){
 	
 		//test to see if connection already exists
 		$sql = "SELECT *
-			FROM p_d_connection D
-			WHERE D.patient_id = ? AND D.hcp_id = ?";
+				FROM p_d_connection D
+				WHERE D.patient_id = ? AND D.hcp_id = ?";
 		$query = $this->db->query($sql, $inputs);
 		
 		if ($this->db->trans_status() === FALSE) {
@@ -271,7 +353,7 @@ class Connections_model extends Model {
 	 *   -1 in case of error in a query
 	 *   -2 if the connection does not exist
 	 *   -3 if the connection was already accepted
-	 *   0  if everything goes fine
+	 *    0 if everything goes fine
 	 * */
 	function accept_doctor_doctor($inputs) {
 		$query = $this->db->query("SELECT * FROM d_d_connection
@@ -301,6 +383,7 @@ class Connections_model extends Model {
 		return 0;
 	}
 
+
 	/**
 	 * Sets the connection patient-doctor to TRUE only if it was not
 	 * already accepted.
@@ -312,7 +395,7 @@ class Connections_model extends Model {
 	 *   -1 in case of error in a query
 	 *   -2 if the connection does not exist
 	 *   -3 if the connection was already accepted
-	 *   0  if everything goes fine
+	 *    0 if everything goes fine
 	 * */
 	function accept_patient_doctor($inputs) {
 		$query = $this->db->query("SELECT * FROM p_d_connection
@@ -349,14 +432,32 @@ class Connections_model extends Model {
 	 * @param $inputs
 	 *   Is of the form: array(A_id, B_id)
 	 * @return
-	 *   Removes the connection from the P_D_Connection table
+	 *  -1 in case of error in a query
+	 *  -2 if the connection does not exist
+	 *   0 if everything goes fine and the connection from the P_D_Connection table is removed
 	 * */
 	function remove_pd_connection($inputs){
 		
-		$sql = "DELETE FROM p_d_connection
-			WHERE (patient_id = ? AND hcp_id = ?) OR (patient_id = ? AND hcp_id = ?)";
+		//test to see if connection exists
+		$sql = "SELECT *
+				FROM p_d_connection
+				WHERE (patient_id = ? AND hcp_id = ?) OR (patient_id = ? AND hcp_id = ?)";
 		$query = $this->db->query($sql, $inputs[0], $inputs[1], $inputs[1], $inputs[0]);
 		
+		if ($this->db->trans_status() === FALSE)
+			return -1;
+			
+		if ($query->num_rows() < 1)
+			return -2;	
+		
+		$sql = "DELETE FROM p_d_connection
+				WHERE (patient_id = ? AND hcp_id = ?) OR (patient_id = ? AND hcp_id = ?)";
+		$query = $this->db->query($sql, $inputs[0], $inputs[1], $inputs[1], $inputs[0]);
+		
+		if ($this->db->trans_status() === FALSE)
+			return -1;
+			
+		return 0;
 	}
 	
 	/**
@@ -365,14 +466,33 @@ class Connections_model extends Model {
 	 * @param $inputs
 	 *   Is of the form: array(A_id, B_id)
 	 * @return
-	 *   Removes the connection from the D_D_Connection table
+	 *  -1 in case of error in a query
+	 *  -2 if the connection does not exist
+	 *   0 if everything goes fine and the connection from the D_D_Connection table is removed
 	 * */
 	function remove_dd_connection($inputs){
-	
-		$sql = "DELETE FROM d_d_connection
-			WHERE (requester_id = ? AND accepter_id = ?) OR (requester_id = ? AND accepter_id = ?)";
+		
+		//test to see if connection exists
+		$sql = "SELECT *
+				FROM d_d_connection
+				WHERE (requester_id = ? AND accepter_id = ?) OR (requester_id = ? AND accepter_id = ?)";
 		$query = $this->db->query($sql, $inputs[0], $inputs[1], $inputs[1], $inputs[0]);
+		
+		if ($this->db->trans_status() === FALSE)
+			return -1;
+			
+		if ($query->num_rows() < 1)
+			return -2;
+		
+		$sql = "DELETE FROM d_d_connection
+				WHERE (requester_id = ? AND accepter_id = ?) OR (requester_id = ? AND accepter_id = ?)";
+		$query = $this->db->query($sql, $inputs[0], $inputs[1], $inputs[1], $inputs[0]);
+		
+		if ($this->db->trans_status() === FALSE)
+			return -1;	
+			
+		return 0;
 	}
 }
-/** @} */
+
 ?>
